posix32-crt: tls.c: check for _DEBUG instead of HAVE_CRTDBG_H
config.h defines HAVE_CRTDBG_H when debug functions declared in
crtdbg.h are present and can be used. This is needed for mingw-w64,
which does not define _DEBUG when compiling against debug version
of CRT.

When HAVE_CRTDBG_H is defined, config-internal.h ensures that
_DEBUG is also defined, which makes checking for HAVE_CRTDBG_H
unnecessary and inconsistent among source files.